+++ litebridge-db/litebridge-db-oracle/src/main/java/module-info.java
@@ -7,13 +7,15 @@ import org.litebridge.db.oracle.convert.OracleOffsetDateTimeConverter;
 @NullMarked
 module litebridge.db.oracle {
     requires org.jspecify;
+    requires java.sql;
+    requires org.slf4j;
+    requires com.oracle.database.jdbc;
+
+    requires litebridge.commons;
     requires litebridge.converter;
     requires litebridge.db.spi;
     requires litebridge.db.spi.impl;
-    requires java.sql;
-    requires org.slf4j;
-    requires litebridge.commons;
-    requires com.oracle.database.jdbc;
+    requires litebridge.orm;
 
     provides org.litebridge.db.spi.DatabaseProvider with org.litebridge.db.oracle.OracleDatabaseProvider;
     provides org.litebridge.convert.converter.Converter with OracleOffsetDateTimeConverter;
@@ -21,4 +23,5 @@ module litebridge.db.oracle {
 
     exports org.litebridge.db.oracle;
     exports org.litebridge.db.oracle.convert;
+    exports org.litebridge.db.oracle.api;
 }